Put utility methods in a single file
